Updates on resume upload

See merge request steffbeckers/cv-matching!10

+++ back-end/RJM.API/ViewModels/DocumentVM.cs
@@ -49,7 +49,7 @@ namespace RJM.API.ViewModels
         /// <summary>
         /// The SizeInBytes property of Document.
         /// </summary>
-        public int? SizeInBytes { get; set; }
+        public long? SizeInBytes { get; set; }
 
         /// <summary>
         /// The FileLastModifiedOn property of Document.
